Add configuration files and setup project dependencies
The changes introduced include creation and configuration of babel and typescript files which are essential for the project setup. The commits also contain the addition of necessary dependencies needed to run the project in the package.json file. The .gitignore file was updated to ignore node_modules and other config files.